Adding infrastructure for coordinate system
Currently nonfunctional.

Worked with Benjamin to figure out a way of crawling the project to get all of the axis.

Current idea is to have the user point to the project that has the axis and get all of them, then populate an array. This would be unnecessary if there was a way of getting the axis contained within the coordinate system.

Must modify the motion axis to be a motion axis or coordinate - consider override