Coordinate Systems - updates and bugfixes
Modified the Motion class to include a boolean for whether or not the module controls a coordinate system. This is instead of using the empty string primitive - makes things easier to read.

Realized I had not added some utilities for the coordinate system. Added Power On. Need to modify Read.
Fixes #19