Example State Machine work
Removed old example state machine (which was created for when the module was outside of DCAF).

Created and started work on a DCAF module as an example state machine.

Cleaned up testing folders - moved the DCAF-created tests from their own folder to a sub-folder within the Testing folder.

Removed redundant PCFG files.

Moved the DumpToLog into Testing as well as the UI that I was using for some of the Runtime examples.