Simple Sequencer, changes to the Utilities
Added functionality to the utilities - the Get Status for Position was built.

Noticed that the command for position profile and the get are swapped as to whether or not the move is relative. Adding a new Issue.

Created code for the Getting Started Experience. Currently am having some issues with nor every move acting.

Fixes #8