Basic state machine work
State machine is now currently functional - but it does not close when done, it will simply wait. (Fixes #28)

May want to change the state machine scheme to avoid dequeueing the first element because on first-call the axis is considered to have "finished" a move.

Additionally, I found that part of the problem for the sequencer was that the motion module's statusword was not correctly updated. This is because the asynchronous axis do not seem to update the read method for the move references the same way as the Synchronous. Therefore I moved to using the Move Done method on the axis itself instead of the move reference.  (Fixes #31)

Also Fixes #29 by making the "true" case on the utilities and the process Relative moves as per ds402.