Performance and Reference Move improvements
Changed Key Value Pairs to Axis or Coord System and the Array of Axis. Changed system to just check if the Axis Array is empty instead of requiring a third item.  Fixes #41

Performance improvements - removed the enable check on the Read (seemed superfluous, can be added to a bulk-read method). Modified the Read to replace the array instead of building it, Fixes #37

Also added to the Reference move the same Skip Read fix that was added to the Position Profile move. It may be possible to skip the Read alltogether if the other values are also from the previous iteration.

Removed superfluous error case structures for performance issues. Additionally, modified the Stop so that it can still be called in the case of an error somewhere else.