Changing Read to Read Fast and Read Full
Previously I was doing the same read in the move modes as well as the Reserved "Read" mode. Now the read has been split into a "fast" read that just gets the position of the coordinate system or axis and velocity (of the full system)  and a "Full" read that was doing what the previous Read did (and a bit more info, now I added the commanded velocity).

Also added a toggle in the controlword to skip the Read for the move modes.

This should greatly increase performance on systems with many axis.