Fixing the Disable Voltage issue
Worked on the switching hte Disable Voltage command sending the system to the wrong state.

I noticed that the CommandToWord was only clearing bit 0 instead of bits 0-3.

Other changes: The ExampleSequencer now passes the current controlword, statusword etc to the Previous statusword in case I need to compare with a previous controlword. Changes the code to only use the current one.