Updates to the Reference Move and Sequencer project
Updated UI for the Sample Sequencer Project and added a version of the project that uses an IC with an AKD.

Using this to test, I found out I was calling the incorrect case for reference moves (same issue as the Stop Moves before).

Was getting a resource error on stop - currently calling a blocking stop before doing a reference move, which will make the loop miss timing when a reference move is called. I will look into how to get that to work, but it seemed acceptable for now.

Also fixes #68