Moved Read and Check Faults outside of the main state
Created Read utility (Fixes #62)

Moved the Fast Read and Full Read outside of the main structure. This allows for reading the position, velocity, IO, etc in any of the non-startup states.

Furthermore, moved the Check Faults logic to within the Full Read. If bit 13 is True, then it will Check Faults for every axis. This is a long process, so it's likely for the system to run late during the iteration.

This fixes #42 - somewhat. I may return to this to optimize for speed.

Between these, this also fixes #46

Finally, also fixes #47