Testing out flag updates
Asynchronous line moves invoke nodes for reading status seem to be incorrect. So I moved back to the Axis properties which will not quite work for Velocity moves.

I noticed that you have to command Reference moves multiple times for them to actually start - the asynchronous express VI seems to be keeping the Execute asserted until the move begins as a workaround, so I will need to do the same.